+++ src/methods/$to-array/__tests__/async-to-array.auto.spec.ts
@@ -6,9 +6,10 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import { asyncToArray, range } from '../../..';
+import { asyncToArray } from '../../..';
+import { asyncRange } from '../../../__tests__/async-range';
 describe('asyncToArray', () => {
-  it('works', async () => {
-    expect(await asyncToArray(range(3))).toEqual([0, 1, 2]);
+  it('turns an iterable into an array', async () => {
+    expect(await asyncToArray(asyncRange(0, 3))).toEqual([0, 1, 2]);
   });
 });