+++ src/methods/$to-array/__tests__/to-array.test.js
@@ -8,9 +8,10 @@
 
 /* eslint-disable no-unused-vars,import/no-duplicates,no-constant-condition */
 
-import { toArray, range } from '../../..';
+import { toArray } from '../../..';
+import { range } from '../../../__tests__/range';
 describe('toArray', () => {
-  it('works', () => {
-    expect(toArray(range(3))).toEqual([0, 1, 2]);
+  it('turns an iterable into an array', () => {
+    expect(toArray(range(0, 3))).toEqual([0, 1, 2]);
   });
 });