+++ src/methods/object-from-async/object-from-async.mjs
@@ -1,0 +1,1 @@
+export { default, asyncToObject as objectFromAsync } from '../$to-object/async-to-object';